solve double first item bug + final code cleanup
On first message socket event, the item was being added twice on the screen, even though it was displayed in the database only once.
The reason for that is that we returned array reference from `getMessages`, which was directly used to set state in the Cubit. For every new message, the array was copied via the ...spread operator. The first time, we spread the array reference that also gets updated inside the channel service, and that resulted in having two instances of the newly arrived messages in the cubit state (but not in the messages DB).